Fix validate script
